+++ tests/test_multiprocessing_serverprocess.py
@@ -2,6 +2,7 @@ import multiprocessing as mp
 import threading
 import time
 from multiprocessing import active_children
+import sys
 
 import pytest
 from mpservice.multiprocessing import (
@@ -217,14 +218,16 @@ def test_shared_memory():
         assert len(mem.buf) == 10
         mem.buf[4] = 100
 
-        print('memory blocks in main:', manager.list_memory_blocks())
-        assert len(manager.list_memory_blocks()) == 1
+        blocks = manager.list_memory_blocks()
+        print('memory blocks in main:', blocks)
+        assert len(blocks) == 1
 
         q = Queue()
         q.put(mem)
         del mem
 
-        assert len(manager.list_memory_blocks()) == 1
+        blocks = manager.list_memory_blocks()
+        assert len(blocks) == 1
 
         p = Process(target=inc_worker, args=(q,))
         p.start()
@@ -243,6 +246,9 @@ def test_shared_memory_from_serverprocess():
         MemoryWorker, method_to_typeid={'memory_block': 'memory_block_in_server'}
     )
     with ServerProcess() as server:
+        print('')
         worker = server.MemoryWorker()
         mem = worker.memory_block(20)
-        print(type(mem))
+        del mem
+        block = server.MemoryBlock(8)
+