+++ CHANGELOG.md
@@ -5,16 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
-## [0.13.0] - in progress
+## [0.13.0] - 2023-05-31
 
 - Breaking changes to ``mpserver.Server`` API: the class is split into two classes: ``Server`` and ``AsyncServer``.
 - ``mpserver.Server.call`` got new parameter ``backpressure`` (previously only the async call has this parameter).
 - Finetuned waiting and sleeping logic in ``mpserver.{Server, AsyncServer}``; use ``Condition`` to replace sleeping.
-- Made sure (or confirmed) that ``mpserver.Server._call`` and ``mpserver.Server._stream` are thread-safe.
+- Made sure (or confirmed) that ``mpserver.Server.call`` and ``mpserver.Server.stream` are thread-safe.
 - ``streamer.Stream.peek`` finetune of printing; got new parameter ``prefix`` and ``suffix``.
 - Refinements to classes ``streamer.{IterQueue, IterProcessQueue, AsyncIterQueue}``.
 - Refinements to ``multiprocessing.ServerProcess``: further diverge from the standard class.
-- New facilities for "shared memory" in the class ``multiprocessing.ServerProcess``.
+- Initial for "shared memory" in the class ``multiprocessing.ServerProcess``.
 
 
 ## [0.12.9] - 2023-05-23